refactor(useKeyPress): rename keyboard_nav, move to @/components
